dynamically imported auction cards to AuctionsGrid and AuctionsList components and added skeleton loading UI
