+++ src/app/ui/auctions/AuctionsGrid.tsx
@@ -1,6 +1,25 @@
 import React, { Suspense } from "react";
 import { TimerProvider } from "@/app/_context/TimerContext";
-import { GamesCard } from "@/app/components/card";
+import dynamic from "next/dynamic";
+
+const DynamicGamesCard = dynamic(() => import("../../components/games_card"), {
+  ssr: false,
+  loading: () => (
+    <div className="tw-flex tw-mt-8 tw-justify-evenly">
+      <div className="tw-flex tw-flex-col tw-m-2">
+        <div className="tw-w-96 tw-mb-5 tw-h-48 tw-bg-gray-700 tw-rounded-md"></div>
+        <div className="tw-w-4/5 tw-h-10 tw-mb-5 tw-bg-gray-700 tw-rounded-lg tw-animate-pulse"></div>
+        <div className="tw-w-full tw-mb-2 tw-h-3 tw-bg-gray-700 tw-rounded-lg tw-animate-pulse"></div>
+        <div className="tw-w-full tw-mb-2 tw-h-3 tw-bg-gray-700 tw-rounded-lg tw-animate-pulse"></div>
+        <div className="tw-w-full tw-mb-2 tw-h-3 tw-bg-gray-700 tw-rounded-lg tw-animate-pulse"></div>
+        <div className="tw-w-full tw-mb-5 tw-h-3 tw-bg-gray-700 tw-rounded-lg tw-animate-pulse"></div>
+        <div className="tw-w-full tw-mb-2 tw-h-3 tw-bg-gray-700 tw-rounded-lg tw-animate-pulse"></div>
+        <div className="tw-w-full tw-mb-10 tw-h-3 tw-bg-gray-700 tw-rounded-lg tw-animate-pulse"></div>
+        <div className="tw-w-1/3 tw-mb-2 tw-h-10 tw-bg-gray-700 tw-rounded-lg tw-animate-pulse"></div>
+      </div>
+    </div>
+  ),
+});
 
 const AuctionsGrid = ({ listing }: { listing: any }) => {
   return (
@@ -11,7 +30,7 @@ const AuctionsGrid = ({ listing }: { listing: any }) => {
             <div key={car._id ? car._id : index + "gamesCard"}>
               <Suspense>
                 <TimerProvider deadline={new Date(car.deadline)}>
-                  <GamesCard
+                  <DynamicGamesCard
                     auction_id={
                       car.auction_id ? car.auction_id : index + "auctionId"
                     }
