+++ src/app/ui/auctions/AuctionsList.tsx
@@ -1,6 +1,24 @@
 import React, { Suspense } from "react";
 import { TimerProvider } from "@/app/_context/TimerContext";
-import { AuctionsListCard, GamesCard } from "@/app/components/card";
+import AuctionsListCard from "@/app/components/auctions_list_card";
+import dynamic from "next/dynamic";
+
+const DynamicAuctionsListCard = dynamic(
+  () => import("../../components/auctions_list_card"),
+  {
+    ssr: false,
+    loading: () => (
+<div className="tw-flex tw-flex-row tw-gap-4">
+  <div className="tw-mb-5 tw-bg-gray-700 tw-max-w-[156px] sm:tw-max-w-[416px] tw-w-full tw-min-w-[156px] tw-h-auto  tw-min-h-[147px] xl:tw-h-[240px] tw-rounded tw-object-cover tw-aspect-auto hover:tw-cursor-pointer"></div>
+  <div className="tw-flex tw-flex-col tw-w-auto tw-flex-grow tw-gap-2">
+    <div className="tw-w-2/3 tw-mb-2 tw-h-8 tw-bg-gray-700 tw-rounded-lg tw-animate-pulse"></div>
+    <div className="tw-w-2/3 tw-mb-2 tw-h-6 tw-bg-gray-700 tw-rounded-lg tw-animate-pulse"></div>
+    <div className="tw-w-2/3 tw-mb-2 tw-h-10 tw-bg-gray-700 tw-rounded-lg tw-animate-pulse"></div>
+  </div>
+</div>
+    ),
+  }
+);
 
 const AuctionsList = ({ listing }: { listing: any }) => {
   return (
@@ -10,7 +28,7 @@ const AuctionsList = ({ listing }: { listing: any }) => {
           <div key={car._id ? car._id : index + "gamesCard"}>
             <Suspense>
               <TimerProvider deadline={new Date(car.deadline)}>
-                <AuctionsListCard
+                <DynamicAuctionsListCard
                   auction_id={
                     car.auction_id ? car.auction_id : index + "auctionId"
                   }
